+++ GoogleBooks/src/components/searchBar/SearchBar.jsx
@@ -1,7 +1,14 @@
 import React from "react";
 
 const SearchBar = () => {
-  return <div>SearchBar</div>;
+  let placeholder = "";
+
+  return (
+    <>
+      <h2>I would like to be a searchbar</h2>
+      {/* <input value={placeholder}>Submit</input> */}
+    </>
+  );
 };
 
 export default SearchBar;
